Refactor advanced ZKP module into separate files with improved structure
